bus-booking-system/requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
Flask-Session==0.5.0